# docs/swagger.yaml
basePath: development
definitions:
...
  request.TenorRequest:
    properties:
      brand:
        type: string
      carCondition:
        type: string
      caroserie:
        type: integer
      city:
        type: string
      fee:
        $ref: '#/definitions/request.fee'
      insurances:
        $ref: '#/definitions/request.insurances'
      loanPackageName:
        type: string
      manufacturedYear:
        type: string
      model:
        type: string
      paymentType:
        type: string
      province:
        type: string
      provisionPercentage:
        type: integer
      simulationType:
        type: string
      simulationValue:
        type: integer
      tenorInMonths:
        type: integer
      unitPrice:
        type: integer
      variant:
        type: string
    type: object
...
  request.fee:
    properties:
      adminFee:
        type: integer
      beaPolis:
        type: integer
    type: object
...
  request.insurances:
    properties:
      additionalInsurances:
        items:
          type: string
        type: array
      insuranceType:
        type: string
      lifeInsurance:
        type: boolean
      putAsOnLoan:
        type: boolean
      tanggungJawabPihakKetiga:
        $ref: '#/definitions/request.tanggungJawabPihakKetiga'
    type: object
...
  request.tanggungJawabPihakKetiga:
    properties:
      isApplied:
        type: boolean
      uangPertanggungan:
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      error_id:
        type: integer
      message:
        additionalProperties:
          type: string
        type: object
    type: object
...
  response.TenorResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: boolean
      statusCode:
        type: string
    type: object
...
host: xuu77ziiri.execute-api.us-east-2.amazonaws.com/
info:
  contact: {}
  description: MMKSI Middleware API Documentation
  title: MMKSI Middleware API Documentation
  version: "1.0"
paths:
...
  /dsf/calculator/allTenors:
    post:
      description: Get Calculator All Tenor
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Application Name
        in: header
        name: applicationName
        required: true
        type: string
      - description: Request Body
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.TenorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TenorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get All Tenor
      tags:
      - Calculator
  /dsf/calculator/perTenor:
    post:
      description: Get Calculator Pertenor
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Application Name
        in: header
        name: applicationName
        required: true
        type: string
      - description: Request Body
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.TenorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TenorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Pertenor
      tags:
      - Calculator
...
schemes:
- https
- http
swagger: "2.0"